move xml utils
